+++ include/aca_grpc.h
@@ -29,6 +29,9 @@ using grpc::Status;
 
 class GoalStateProvisionerImpl final : public GoalStateProvisioner::Service {
   public:
+  std::unique_ptr<GoalStateProvisioner::Stub> stub_;
+  HostRequestReply RequestGoalStates(HostRequest *request);
+  
   // ~GoalStateProvisionerImpl();
   explicit GoalStateProvisionerImpl(){};
   Status PushNetworkResourceStates(ServerContext *context, const GoalState *goalState,
