+++ src/ovs/ovs_control.cpp
@@ -16,7 +16,7 @@
 #include "aca_log.h"
 #include "aca_util.h"
 #include "ovs_control.h"
-#include "aca_ovs_control.h"
+#include "aca_on_demand_engine.h"
 #include <sstream> // std::(istringstream)
 #include <string> // std::(string)
 #include <signal.h>
@@ -41,7 +41,7 @@
 #include <openvswitch/vlog.h>
 
 using namespace std;
-using namespace aca_ovs_control;
+using namespace aca_on_demand_engine;
 
 extern std::atomic_ulong g_total_execute_openflow_time;
 
@@ -79,9 +79,9 @@ char *ds_steal_cstr(struct ds *);
 VLOG_DEFINE_THIS_MODULE(ovs_control)
 
 namespace ovs_control
-{
+{ 
 OVS_Control &OVS_Control::get_instance()
-{
+{ 
   // Instance is destroyed when program exits.
   // It is instantiated on first use.
   static OVS_Control instance;
@@ -719,7 +719,7 @@ const char *OVS_Control::openflow_from_hex(const char *hex, ofpbuf **msgp)
  * NXT_PACKET_IN2 that includes a continuation. */
 void OVS_Control::monitor_vconn(vconn *vconn, bool reply_to_echo_requests,
                                 bool resume_continuations, const char *bridge_)
-{
+{ 
   static const char *bridge = bridge_;
   bool timestamp = true;
   struct barrier_aux barrier_aux = { vconn, NULL };
@@ -991,7 +991,7 @@ void OVS_Control::monitor_vconn(vconn *vconn, bool reply_to_echo_requests,
           error = ofputil_decode_packet_in((ofp_header *)b->data, true, NULL, NULL,
                                            &pin, &total_lenp, &buffer_idp, &continuation);
           uint32_t in_port = pin.flow_metadata.flow.in_port.ofp_port;
-          ACA_OVS_Control::get_instance().parse_packet(in_port, pin.packet);
+          ACA_On_Demand_Engine::get_instance().parse_packet(in_port, pin.packet);
 
           if (error) {
             fprintf(stderr, "decoding packet-in failed: %s",
